+++ internal/api/grpc_server.go
@@ -26,6 +26,7 @@ func StartGrpcServer(logger zerolog.Logger, s *config.Settings, m mediator.Media
 	deviceDefinitionService := NewGrpcService(m, &logger)
 	recallsService := NewGrpcRecallsService(m, &logger)
 	reviewsService := NewGrpcReviewsService(m, &logger)
+	integrationService := NewGrpcIntegrationService(m, &logger)
 
 	server := grpc.NewServer(
 		grpc.UnaryInterceptor(grpc_middleware.ChainUnaryServer(
@@ -36,6 +37,7 @@ func StartGrpcServer(logger zerolog.Logger, s *config.Settings, m mediator.Media
 	pkggrpc.RegisterDeviceDefinitionServiceServer(server, deviceDefinitionService)
 	pkggrpc.RegisterRecallsServiceServer(server, recallsService)
 	pkggrpc.RegisterReviewsServiceServer(server, reviewsService)
+	pkggrpc.RegisterIntegrationServiceServer(server, integrationService)
 
 	logger.Info().Str("port", s.GRPCPort).Msgf("started grpc server on port: %v", s.GRPCPort)
