+++ pkg/grpc/device_definition.proto
@@ -275,33 +275,6 @@ message GetDeviceStyleByIDRequest {
   string id = 1;
 }
 
-message GetDeviceCompatibilityListRequest {
-  string make_id = 1;
-  string integration_id = 2;
-  string region = 3;
-}
-
-message Feature {
-  string key = 1;
-  int32 support_level = 2;
-  string css_icon = 3;
-}
-
-message DeviceCompatibilities {
-  int32 year = 1;
-  repeated Feature features = 2;
-  string level = 3;
-}
-
-message DeviceCompatibilityList {
-  string name = 1;
-  repeated DeviceCompatibilities years = 2;
-}
-
-message GetDeviceCompatibilityListResponse {
-  repeated DeviceCompatibilityList models = 1;
-}
-
 message GetDeviceStyleByDeviceDefinitionIDRequest {
   string id = 1;
 }
@@ -393,34 +366,6 @@ message DeleteDeviceTypeRequest {
   string id = 1;
 }
 
-message GetIntegrationFeatureByIDRequest {
-  string id = 1;
-}
-
-message CreateOrUpdateIntegrationFeatureRequest {
-  string id = 1;
-  string elastic_property = 2;
-  string display_name = 3;
-  string css_icon = 4;
-  float feature_weight = 5;
-}
-
-message DeleteIntegrationFeatureRequest {
-  string id = 1;
-}
-
-message GetIntegrationFeatureListResponse {
-  repeated GetIntegrationFeatureResponse integration_features = 2;
-}
-
-message GetIntegrationFeatureResponse {
-  string feature_key = 1;
-  string elastic_property = 2;
-  string display_name = 3;
-  string css_icon = 4;
-  float feature_weight = 5;
-}
-
 service DeviceDefinitionService {
   rpc GetDeviceDefinitionByID(GetDeviceDefinitionRequest) returns (GetDeviceDefinitionResponse);
   rpc GetDeviceDefinitionBySlug(GetDeviceDefinitionBySlugRequest) returns (GetDeviceDefinitionItemResponse);
@@ -443,7 +388,6 @@ service DeviceDefinitionService {
   rpc UpdateDeviceDefinition(UpdateDeviceDefinitionRequest) returns (BaseResponse);
   rpc SetDeviceDefinitionImage(UpdateDeviceDefinitionImageRequest) returns (BaseResponse);
   rpc GetFilteredDeviceDefinition(FilterDeviceDefinitionRequest) returns (GetFilteredDeviceDefinitionsResponse);
-  rpc GetDeviceCompatibilities(GetDeviceCompatibilityListRequest) returns (GetDeviceCompatibilityListResponse);
   rpc GetDeviceStyleByID(GetDeviceStyleByIDRequest) returns (DeviceStyle);
   rpc GetDeviceStyleByExternalID(GetDeviceStyleByIDRequest) returns (DeviceStyle);
   rpc GetDeviceStylesByDeviceDefinitionID(GetDeviceStyleByDeviceDefinitionIDRequest) returns (GetDeviceStyleResponse);
@@ -456,9 +400,4 @@ service DeviceDefinitionService {
   rpc UpdateDeviceType(UpdateDeviceTypeRequest) returns (BaseResponse);
   rpc DeleteDeviceType(DeleteDeviceTypeRequest) returns (BaseResponse);
 
-  rpc GetIntegrationFeatureByID(GetIntegrationFeatureByIDRequest) returns (GetIntegrationFeatureResponse);
-  rpc GetIntegrationFeatures(google.protobuf.Empty) returns (GetIntegrationFeatureListResponse);
-  rpc CreateIntegrationFeature(CreateOrUpdateIntegrationFeatureRequest) returns (BaseResponse);
-  rpc UpdateIntegrationFeature(CreateOrUpdateIntegrationFeatureRequest) returns (BaseResponse);
-  rpc DeleteIntegrationFeature(DeleteIntegrationFeatureRequest) returns (BaseResponse);
 }
