+++ pkg/grpc/integration.proto
@@ -1,0 +1,75 @@
+syntax = "proto3";
+
+import "google/protobuf/empty.proto";
+
+option go_package = "github.com/DIMO-Network/device-definitions-api/pkg/grpc";
+
+package grpc;
+
+message IntegrationBaseResponse {
+  string id = 1;
+}
+
+message GetIntegrationFeatureByIDRequest {
+  string id = 1;
+}
+
+message CreateOrUpdateIntegrationFeatureRequest {
+  string id = 1;
+  string elastic_property = 2;
+  string display_name = 3;
+  string css_icon = 4;
+  float feature_weight = 5;
+}
+
+message DeleteIntegrationFeatureRequest {
+  string id = 1;
+}
+
+message GetIntegrationFeatureListResponse {
+  repeated GetIntegrationFeatureResponse integration_features = 2;
+}
+
+message GetIntegrationFeatureResponse {
+  string feature_key = 1;
+  string elastic_property = 2;
+  string display_name = 3;
+  string css_icon = 4;
+  float feature_weight = 5;
+}
+
+message GetDeviceCompatibilityListRequest {
+  string make_id = 1;
+  string integration_id = 2;
+  string region = 3;
+}
+
+message GetDeviceCompatibilityListResponse {
+  repeated DeviceCompatibilityList models = 1;
+}
+
+message DeviceCompatibilityList {
+  string name = 1;
+  repeated DeviceCompatibilities years = 2;
+}
+
+message DeviceCompatibilities {
+  int32 year = 1;
+  repeated Feature features = 2;
+  string level = 3;
+}
+
+message Feature {
+  string key = 1;
+  int32 support_level = 2;
+  string css_icon = 3;
+}
+
+service IntegrationService {
+  rpc GetDeviceCompatibilities(GetDeviceCompatibilityListRequest) returns (GetDeviceCompatibilityListResponse);
+  rpc GetIntegrationFeatureByID(GetIntegrationFeatureByIDRequest) returns (GetIntegrationFeatureResponse);
+  rpc GetIntegrationFeatures(google.protobuf.Empty) returns (GetIntegrationFeatureListResponse);
+  rpc CreateIntegrationFeature(CreateOrUpdateIntegrationFeatureRequest) returns (IntegrationBaseResponse);
+  rpc UpdateIntegrationFeature(CreateOrUpdateIntegrationFeatureRequest) returns (IntegrationBaseResponse);
+  rpc DeleteIntegrationFeature(DeleteIntegrationFeatureRequest) returns (IntegrationBaseResponse);
+}
